Upgrade css-select
A dependency of css-select for versions below 4.1.3 is flagged by
various audit tools such as npm audit and OWASP Dependency Check. Update
so this package is no longer flagged.

From the description of breaking changes in v4 of css-select[^1] and the
current usage of it this should be a non-breaking change for this
project (some pseudo selectors are stricter, HTML attributes made case
insensitive, `strict` option removed).

[^1]: https://github.com/fb55/css-select/releases/tag/v4.0.0